agent/requirements.txt
crewai
langchain-openai
python-dotenv
langchain-community
